requirements.txt
flask
numpy
pandas
tensorflow
Pillow
scikit-learn
opencv-python
matplotlib
streamlit
